+++ src/components/avatar/avatar.ts
@@ -1,4 +1,4 @@
-import { Component, prepareComponent } from '../../core/base/component';
+import { Component, prepareComponent } from '../../core/component';
 import './avatar.scss';
 
 export type TAvatarProps = {
