--- src/core/base/component/index.ts
@@ -1,1 +1,0 @@
-export * from './Component';
