--- src/core/base/models/enums.ts
@@ -1,11 +1,0 @@
-export enum EEvents {
-  INIT = 'INIT',
-  COMPILE = 'COMPILE',
-  MOUNT = 'COMPONENT_DID_MOUNT',
-  UPDATE = 'COMPONENT_DID_UPDATE',
-}
-
-export enum EManageEventsAction {
-  ADD = 'ADD',
-  REMOVE = 'REMOVE',
-}
