+++ src/core/component/enhancers/childrenController/index.ts
