+++ src/core/component/enhancers/index.ts
@@ -1,0 +1,2 @@
+export * from './childrenController';
+export * from './elementController';
