+++ src/core/component/index.ts
@@ -1,0 +1,2 @@
+export * from './Component';
+export { renderDOM } from './utils';
