+++ src/core/component/models/enums.ts
@@ -1,0 +1,4 @@
+export enum EManageEventsAction {
+  ADD = 'ADD',
+  REMOVE = 'REMOVE',
+}
