+++ src/core/component/models/index.ts
