+++ src/core/httpClient/utils/index.ts
@@ -1,6 +1,6 @@
-import { TSimpleObject } from '../models';
+import { ISimpleObject } from '../../models';
 
-export function queryStringify(data?: TSimpleObject) {
+export function queryStringify(data?: ISimpleObject) {
   if (!data) return '';
   const queryParams = Object.keys(data).map((key) => `${key}=${data[key]}`);
   return `?${queryParams.join('&')}`;
