+++ src/core/models/index.ts
@@ -1,2 +1,3 @@
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type AnyType = any;
+export type ISimpleObject = Record<string | symbol | number, AnyType>;
