+++ src/pages/signin-page/signin-page.ts
@@ -1,11 +1,11 @@
-import { prepareComponent } from '../../core/base/component';
 import './signin-page.scss';
 import { Form } from '../../components/form/form';
 import { validatePassword, validateLogin, validateEmail, validateName, validatePhone } from '../../utils';
 import { TInputFieldProps } from '../../ui-kit/input-field';
 import { TButtonProps } from '../../ui-kit/button';
 import { TLinkProps } from '../../ui-kit/link';
-import { ISimpleObject } from '../../core/base/models';
+import { prepareComponent } from '../../core/component';
+import { ISimpleObject } from '../../core/models';
 
 type TSigninPageState = {
   fields: TInputFieldProps[];
