+++ src/ui-kit/button/button.ts
@@ -1,4 +1,4 @@
-import { Component, prepareComponent } from '../../core/base/component';
+import { Component, prepareComponent } from '../../core/component';
 import './button.scss';
 
 export type TButtonProps = {
