+++ src/ui-kit/file-field/file-field.ts
@@ -1,4 +1,4 @@
-import { Component, prepareComponent } from '../../core/base/component';
+import { Component, prepareComponent } from '../../core/component';
 import './file-field.scss';
 
 export type TFileFieldProps = {
