+++ src/ui-kit/input-field/input-field.ts
@@ -1,4 +1,4 @@
-import { Component, prepareComponent } from '../../core/base/component';
+import { Component, prepareComponent } from '../../core/component';
 import './input-field.scss';
 
 export type TInputFieldProps = {
