+++ src/ui-kit/search-input/search-input.ts
@@ -1,4 +1,4 @@
-import { Component, prepareComponent } from '../../core/base/component';
+import { Component, prepareComponent } from '../../core/component';
 import './search-input.scss';
 
 type TSearchInputProps = {
