+++ src/ui-kit/text-button/text-button.ts
@@ -1,4 +1,4 @@
-import { Component, prepareComponent } from '../../core/base/component';
+import { Component, prepareComponent } from '../../core/component';
 import './text-button.scss';
 
 export type TTextButtonProps = {
